renamed ServiceProvider models folder to serviceProviderInfo folder
