+++ HomeServiceTracker/Shared/Models/ServiceProviderInfo/ServiceProviderInfoCreate.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace HomeServiceTracker.Shared.Models.ServiceProviderInfo
+{
+    public class ServiceProviderInfoCreate
+    {
+        [Required]
+        public string ServiceProviderName { get; set; }
+    }
+}
