+++ HomeServiceTracker/Shared/Models/ServiceProviderInfo/ServiceProviderInfoDetail.cs
@@ -1,17 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HomeServiceTracker.Shared.Models.ServiceProvider
+namespace HomeServiceTracker.Shared.Models.ServiceProviderInfo
 {
-    public class ServiceProviderEdit
+    public class ServiceProviderInfoDetail
     {
-        [Required]
         public int Id { get; set; }
-        [Required]
         public string ServiceProviderName { get; set; }
+        public int NumberOfServices { get; set; }
+        public float AverageServiceProviderRating { get; set; }
     }
 }
