+++ HomeServiceTracker/Shared/Models/ServiceProviderInfo/ServiceProviderInfoEdit.cs
@@ -5,10 +5,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HomeServiceTracker.Shared.Models.ServiceProvider
+namespace HomeServiceTracker.Shared.Models.ServiceProviderInfo
 {
-    public class ServiceProviderCreate
+    public class ServiceProviderInfoEdit
     {
+        [Required]
+        public int Id { get; set; }
         [Required]
         public string ServiceProviderName { get; set; }
     }
